# api/openapi.yaml
openapi: "3.0.3"
info:
  title: Spec Doc API
  version: "1.0.0"
  description: >
    Phase 1 contract — health check, projects CRUD, and context
    read/write for four named keys (builder, principles, codebase,
    references). This file is the build artifact; flask/dtos/ and
    mock_server.py are derived from it.
servers:
  - url: http://localhost:3101
    description: Flask development server
  - url: http://localhost:3102
    description: Mock server (Task 3)
paths:
...
  /api/ai/text/rewrite:
    post:
      summary: Rewrite text with optional instructions
      operationId: rewriteText
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RewriteRequest'
      responses:
        '200':
          description: Rewritten text
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RewriteResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '502':
          description: Upstream AI provider failure
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/ai/text/generate-spec:
    post:
      summary: Generate specification documents from a brain dump
      operationId: generateSpec
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerateSpecRequest'
      responses:
        '200':
          description: "Raw multi-file spec text with ===FILE: filename=== markers"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenerateSpecResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '502':
          description: Upstream AI provider failure
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/ai/text/iterate:
    post:
      summary: Iterate on a document with an instruction
      operationId: iterateText
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IterateRequest'
      responses:
        '200':
          description: Iterated document text
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IterateResponse'
        '400':
          description: document is empty after whitespace stripping
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Request validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '502':
          description: AI provider error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/ai/text/lint-braindump:
    post:
      summary: Analyse a brain dump for spec readiness
      operationId: lintBraindump
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LintBraindumpRequest'
      responses:
        '200':
          description: Readiness assessment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LintBraindumpResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '502':
          description: Upstream AI provider failure or JSON parse failure
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/ai/text/generate:
    post:
      summary: Generate text from a prompt with optional tone, builder-context-injected
      operationId: generateText
      tags: [ai]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerateRequest'
      responses:
        '200':
          description: Generated text
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenerateResponse'
        '400':
          description: prompt is empty after whitespace stripping
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Request validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '502':
          description: AI provider error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/ai/text/review:
    post:
      summary: Review spec documents and return quality scores
      operationId: reviewDocuments
      tags: [ai]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewRequest'
      responses:
        '200':
          description: Review scores and issues
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '502':
          description: AI provider error or JSON parse failure
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/ai/text/bootstrap-project:
    post:
      summary: >
        Bootstrap a full capability folder — analysis → epic → architecture chained
        server-side. Returns 202 with a job_id; poll the status endpoint for the result.
      operationId: bootstrapProject
      tags: [ai]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BootstrapProjectRequest'
      responses:
        '202':
          description: Job accepted; poll GET /bootstrap-project/status/{job_id} for the result.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BootstrapStartResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          $ref: '#/components/responses/UsageLimitReached'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/ai/text/bootstrap-project/status/{job_id}:
    get:
      summary: Poll the status of a bootstrap job.
      operationId: getBootstrapProjectStatus
      tags: [ai]
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: string
          description: >
            Job identifier returned by POST /api/ai/text/bootstrap-project.
            The job is evicted from the in-process registry on the first
            terminal (done=true) read; subsequent polls return 404.
      responses:
        '200':
          description: Current status of the bootstrap job.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BootstrapJobStatus'
        '404':
          description: Job not found or already evicted after first terminal read.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/ai/text/bootstrap-project/{job_id}/cancel:
    post:
      summary: Request cooperative cancellation of an in-flight bootstrap job
      operationId: cancelBootstrap
      tags: [ai]
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '202':
          description: Cancellation requested; status flips to CANCELLING.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CancelAcceptedResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Job is not in a cancellable state.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/ai/text/bootstrap-project/{job_id}/retry:
    post:
      summary: Retry a single bootstrap step (analysis | epic | architecture)
      operationId: retryBootstrap
      tags: [ai]
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RetryBootstrapRequest'
      responses:
        '202':
          description: New job started for the targeted step.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BootstrapStartResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '503':
          description: The targeted sub-workflow is not registered (Rel-T3 lane).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    LintBraindumpRequest:
      type: object
      required: [braindump]
      properties:
        braindump:
          type: string
          minLength: 1
          description: Brain dump text to analyse for spec readiness.
    LintBraindumpResponse:
      type: object
      required: [ready, flags]
      properties:
        ready:
          type: boolean
        flags:
          type: array
          items:
            $ref: '#/components/schemas/LintBraindumpFlag'
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    GenerateRequest:
      type: object
      required: [prompt]
      properties:
        prompt:
          type: string
          minLength: 1
          description: >
            Prompt text to generate from. Whitespace-only values are
            rejected by the server.
        tone:
          type: string
          default: 'balanced'
          description: >
            Optional tone directive (e.g. "balanced", "formal", "casual").
            Defaults to "balanced" to match the Angular client default.
    GenerateResponse:
      type: object
      required: [text, latencyMs]
      properties:
        text:
          type: string
        latencyMs:
          type: integer
          minimum: 0
    RewriteRequest:
      type: object
      required: [text]
      properties:
        text:
          type: string
          minLength: 1
          description: Text to rewrite. Whitespace-only values are rejected by the server.
        instructions:
          type: string
          default: ''
          description: Optional rewrite instruction. Empty string means clean rewrite.
    RewriteResponse:
      type: object
      required: [text, latencyMs]
      properties:
        text:
          type: string
        latencyMs:
          type: integer
          minimum: 0
    GenerateSpecRequest:
      type: object
      required: [input]
      properties:
        input:
          type: string
          minLength: 1
          description: Brain dump text. Whitespace-only values are rejected by the server.
    GenerateSpecResponse:
      type: object
      required: [text, latencyMs]
      properties:
        text:
          type: string
          description: >
            Raw multi-file output with ===FILE: filename=== markers. Not JSON-wrapped.
            The Angular parser splits on this exact string; wrapping breaks the parser.
        latencyMs:
          type: integer
          minimum: 0
    IterateRequest:
      type: object
      required:
        - document
      properties:
        document:
          type: string
          minLength: 1
          description: >
            Document content to iterate on. Whitespace-only values are
            rejected by the server.
        instruction:
          type: string
          default: ''
          description: >
            Optional editing instruction. Empty string means clean/improve
            with no specific directive.
    IterateResponse:
      type: object
      required:
        - text
        - latencyMs
      properties:
        text:
          type: string
        latencyMs:
          type: integer
          minimum: 0
    ReviewRequest:
      type: object
      required: [documents]
      properties:
        documents:
          type: object
          additionalProperties:
            type: string
          description: Map of document name to content.
    ReviewResponse:
      type: object
      required: [scores, issues]
      properties:
        scores:
          type: object
          additionalProperties:
            type: number
          description: Quality scores keyed by dimension (clarity, completeness, etc.)
        issues:
          type: array
          items:
            type: string
          description: List of issues found in the documents.
    BootstrapProjectRequest:
      type: object
      required: [project_name, braindump]
      properties:
        project_name:
          type: string
          minLength: 1
        braindump:
          type: string
          minLength: 1
        builder:
          type: string
          default: ''
        principles:
          type: string
          default: ''
        codebase:
          type: string
          default: ''
        references:
          type: string
          default: ''
...
    BootstrapStartResponse:
      type: object
      description: Immediate 202 acknowledgement from POST /api/ai/text/bootstrap-project.
      required: [job_id]
      properties:
        job_id:
          type: string
          description: >
            Opaque job identifier. Pass to GET /bootstrap-project/status/{job_id} to
            poll for completion. The job is evicted from the server on the first
            terminal (done=true) read.
    BootstrapJobStatus:
      type: object
      description: >
        Status snapshot returned by GET /bootstrap-project/status/{job_id}. Mirrors the
        GenerateTaskStatusResponse shape for Angular polling parity. The 'files',
        'error', and 'latencyMs' fields are present only when the job has reached a
        terminal state.
      required: [running, done]
      properties:
        running:
          type: boolean
          description: True while the background chain is in progress.
        done:
          type: boolean
          description: True once the job has terminated (successfully or not).
        current_step:
          type: string
          nullable: true
          description: >
            Name of the workflow step currently being driven by the runtime,
            or null between iterations and after the workflow completes.
        partial:
          type: string
          description: >
            Rolling 500-char tail of the in-flight step's streamed output.
            Empty string when the current step does not stream or has not
            yet produced any chunks. Read from
            execution.outputs["_partials"][current_step].
        warnings:
          type: array
          items:
            type: string
          description: >
            Per-execution warnings accumulated by the truncation heuristic
            after each successful step. Empty array means clean output.
            The status route always populates this; consumers may rely on it.
        status:
          type: string
          description: >
            Terminal status indicator, present only on terminal CANCELLED reads.
            Successful and errored reads do not populate this field.
        files:
          type: array
          items:
            $ref: '#/components/schemas/BootstrapFile'
          description: Output files produced by the bootstrap run. Present only on success.
        error:
          type: string
          description: Error message captured from the background job. Present only on failure.
        latencyMs:
          type: integer
          minimum: 0
          description: Wall-clock milliseconds from job start to completion.
    CancelAcceptedResponse:
      type: object
      description: >
        Returned with HTTP 202 by the cancel routes once the
        WorkflowExecution has been flipped to CANCELLING. Final
        CANCELLED transition happens between step boundaries on the
        runtime's next iteration.
      required: [status]
      properties:
        status:
          type: string
          description: Current execution status string after the request.
          example: CANCELLING
    RetryBootstrapRequest:
      type: object
      required: [step]
      properties:
        step:
          type: string
          enum: [analysis, epic, architecture]
          description: >
            Bootstrap sub-step to retry. Each maps to a Rel-T3 sub-workflow
            (spec_gen/bootstrap-{step}-only) that re-uses the prior run's
            upstream outputs as inputs.
...
  responses:
    BadRequest:
      description: Missing or invalid request fields
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    UsageLimitReached:
      description: >
        Daily free-tier limit reached for the gated feature. Free-tier caps are
        bootstrap=3 / task_gen=20 / spec_gen=10 per UTC day. Pro users never
        receive this response.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/UsageLimitError"
          example:
            error: free_tier_limit_reached
            feature: task_gen
            limit: 20
            reset_at: "2026-04-27T00:00:00Z"
            upgrade_url: /upgrade
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: >
        Neon Auth JWT issued at login. Pass as Authorization: Bearer <token>.
        Verified by the require_auth decorator (implemented in Task 2).